move files to own components
